refactor: extract auth handlers
